Move pushQueryToSlack to a simple function
